# contract/api/agent-api.openapi.yaml
openapi: 3.1.0
info:
  title: cards402 Agent API
  version: 0.1.0
  description: |
    Virtual Visa cards for AI agents. Pay with USDC or XLM on Stellar,
    receive a card number, CVV, and expiry within ~40 seconds.

    **Authentication:** Every request requires an `X-Api-Key` header with
    a valid API key obtained from the cards402 dashboard.

    **Idempotency:** `POST /orders` accepts an `Idempotency-Key` header.
    Duplicate requests within 24h return the cached response.
  contact:
    name: cards402
    url: https://cards402.com
  license:
    name: Proprietary
servers:
  - url: https://api.cards402.com/v1
    description: Production
  - url: http://localhost:4000/v1
    description: Local development
security:
  - apiKey: []
paths:
  /orders:
    post:
      operationId: createOrder
      summary: Create a new card order
      description: |
        Creates a pending order and returns Soroban contract payment
        instructions. The agent pays the contract on Stellar, the watcher
        detects the event, and fulfillment begins automatically.

        For amounts above the policy approval threshold, returns 202 with
        `phase: awaiting_approval` instead of payment instructions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            maxLength: 128
          description: Deduplication key — same key + same body = cached response for 24h.
      responses:
        '201':
          description: Order created, payment pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreatedResponse'
        '202':
          description: Order requires owner approval before payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAwaitingApprovalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Idempotency-Key reused with different body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Fulfillment system temporarily unavailable (frozen)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      operationId: listOrders
      summary: List this agent's orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              [
                pending_payment,
                ordering,
                delivered,
                failed,
                refunded,
                refund_pending,
                awaiting_approval,
                expired,
                rejected,
              ]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: since_created_at
          in: query
          schema:
            type: string
            format: date-time
          description: Only orders created at or after this timestamp.
        - name: since_updated_at
          in: query
          schema:
            type: string
            format: date-time
          description: Only orders updated at or after this timestamp.
      responses:
        '200':
          description: Array of order summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderListItem'
...
  /policy/check:
    get:
      operationId: checkPolicy
      summary: Dry-run policy check without creating an order
      description: |
        Returns what would happen if the agent tried to create an order
        for the given amount. Useful for pre-flight checks before
        committing to a Stellar transaction.
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
            exclusiveMinimum: 0
      responses:
        '200':
          description: Policy evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyCheckResult'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key from the cards402 dashboard
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - amount_usdc
      properties:
        amount_usdc:
          type: string
          pattern: '^\d+(\.\d+)?$'
          description: Card value in USD as a decimal string, e.g. "10.00"
          examples: ['10.00', '25.50', '100.00']
        webhook_url:
          type: string
          format: uri
          description: HTTPS URL to receive delivery/failure webhooks
        metadata:
          type: object
          description: Arbitrary JSON metadata stored with the order
          additionalProperties: true
    PaymentInstructions:
      type: object
      required: [type, contract_id, order_id, usdc]
      properties:
        type:
          type: string
          enum: [soroban_contract]
        contract_id:
          type: string
          description: Soroban contract address (C...)
        order_id:
          type: string
          format: uuid
          description: Pass this as the order_id arg to pay_usdc/pay_xlm
        usdc:
          type: object
          required: [amount, asset]
          properties:
            amount:
              type: string
              description: USDC amount as 7-decimal string
            asset:
              type: string
              description: 'CODE:ISSUER format, e.g. USDC:GA5Z...'
        xlm:
          type: object
          properties:
            amount:
              type: string
              description: Equivalent XLM amount (spot conversion)
    Budget:
      type: object
      required: [spent_usdc]
      properties:
        spent_usdc:
          type: string
          description: Total spent by this API key
        limit_usdc:
          type: string
          nullable: true
          description: Spend limit (null = unlimited)
        remaining_usdc:
          type: string
          nullable: true
          description: Remaining budget (null = unlimited)
...
    OrderCreatedResponse:
      type: object
      required: [order_id, status, payment, poll_url, budget]
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending_payment]
        phase:
          type: string
          enum: [awaiting_payment]
        amount_usdc:
          type: string
        payment:
          $ref: '#/components/schemas/PaymentInstructions'
        poll_url:
          type: string
          description: Relative URL to poll for status
        budget:
          $ref: '#/components/schemas/Budget'
    OrderAwaitingApprovalResponse:
      type: object
      required: [order_id, phase, approval_request_id, amount_usdc]
      properties:
        order_id:
          type: string
          format: uuid
        phase:
          type: string
          enum: [awaiting_approval]
        approval_request_id:
          type: string
          format: uuid
        amount_usdc:
          type: string
        message:
          type: string
          description: Human-readable reason the order requires approval
        note:
          type: string
          description: Instruction for the agent
        expires_at:
          type: string
          format: date-time
...
    OrderListItem:
      type: object
      required: [id, status, amount_usdc, payment_asset, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        amount_usdc:
          type: string
        payment_asset:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PolicyCheckResult:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, blocked, pending_approval]
        rule:
          type: string
          nullable: true
          description: Which policy rule matched
        reason:
          type: string
          nullable: true
          description: Human-readable explanation
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples: [invalid_amount, rate_limit_exceeded, service_temporarily_unavailable]
        message:
          type: string
          description: Human-readable explanation
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'